+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-    "python.pythonPath": "/Library/Developer/CommandLineTools/usr/bin/python3"
+    "python.pythonPath": "/home/htinoco/.cache/pypoetry/virtualenvs/ttp-sros-parser-JRXJ3YFb-py3.8/bin/python"
 }
