+++ ttp_sros_parser/templates/admin_display_file/sros_router_interface.ttp
@@ -2,17 +2,14 @@
     router {{ignore}}
     <group name="interfaces">
         interface {{interface_name | strip('"')}}
-            shutdown {{admin_state | set(false)}}
-            description {{description | strip('"')}}
+            shutdown {{admin_state | set(false) }}
             description {{description | strip('"')}}
             enable-ingress-stats {{ingress_stats | set(true)}}
-            <group name="ipv4">
-            <group name="primary">
+            <group name="ipv4.primary">
             address {{address}}/{{prefix_length}}
             </group>
-            <group name="bfd">
+            <group name="ipv4.bfd">
             bfd {{ipv4_bfd | re(r"\d.*")}}
-            </group>
             </group>
             ldp_sync_timer {{ldp_sync_timer}}
             cpu_protection {{cpu_protection}}
@@ -35,8 +32,5 @@
             no shutdown {{admin_state | set(true)}}
         exit {{_end_}}
     </group>
-    #-------------------------------------------------- {{_end_}}
 </group>
-        <group name="router">
-        autonomous-system {{autonomous-system}}
-        </group>
+    #-------------------------------------------------- {{_end_}}
